@@ -1,9 +1,10 @@
 import { Rubro } from './rubro';
+import { Producto } from './producto';
 
 export class ArbolApropiacion {
     Rubro: Rubro;
     Vigencia: number;
-    ApropiacionInicial: number;
+    ValorInicial: number;
     Estado: string;
     ApropiacionAnterior: number;
     Codigo: string;
@@ -12,4 +13,5 @@ export class ArbolApropiacion {
     UnidadEjecutora: string;
     Padre: string;
     Hijos: Array<string>;
+    Productos: Array<Producto>;
 }
