@@ -29,14 +29,20 @@ export class ArbolHelper {
       *  returns full rubro's tree information (all nodes and branches).
       * @returns  data with tree structure for the ndTree module.
       */
-    public getFullArbolEstado(vigencia = '0', estado = 'registrada') {
+    public getFullArbolEstado(vigencia = '0', estado = 'registrada', params?: any) {
+
+    let query = '';
+    if (params) {
+      let queryString = Object.keys(params).map(key => key + ':' + params[key]).join(',');
+        query = `?query=${queryString}`;
+    }
         this.rqManager.setPath('PLAN_CUENTAS_MONGO_SERVICE');
         // this.rqManager.setPath('DUMMY_SERVICE');
         // Set the optional branch for the API request.
         const unidadEjecutora = 1;
         // const raiz = 3;
         // call request manager for the tree's data.
-        return this.rqManager.get(`arbol_rubro_apropiacion/arbol_por_estado/${unidadEjecutora}/${vigencia}/${estado}`);
+        return this.rqManager.get(`arbol_rubro_apropiacion/arbol_por_estado/${unidadEjecutora}/${vigencia}/${estado}/${query}`);
     }
 
 }
