@@ -14,6 +14,7 @@ import { RequestManager } from '../../../../@core/managers/requestManager';
 export class ListEntityComponent implements OnInit {
   // Local Inputs ...
   @Input('uuidReadFieldName') uuidReadField: string;
+  @Input('paramsFieldsName') paramsFieldsName: object;
   @Input('uuidDeleteFieldName') uuidDeleteField: string;
   @Input('listColumns') listColumns: object;
 
@@ -31,24 +32,31 @@ export class ListEntityComponent implements OnInit {
   @Input('createConfirmMessage') createConfirmMessage: string;
   @Input('isOnlyCrud') isOnlyCrud: boolean;
   @Input('listSettings') listSettings: object;
-
+  @Input('externalCreate') externalCreate: boolean;
+  @Input('viewItemSelected') viewItemSelected: boolean;
   @Input('loadFormDataFunction') loadFormData: (...params) => Observable<any>;
   @Input('updateEntityFunction') updateEntityFunction: (...params) => Observable<any>;
   @Input('createEntityFunction') createEntityFunction: (...params) => Observable<any>;
   @Output() auxcambiotab = new EventEmitter<boolean>();
+  @Output() crudcambiotab = new EventEmitter<boolean>();
+  @Output() externalTabActivator = new EventEmitter<string>();
+  @Output() infooutput = new EventEmitter<any>();
+  externalTabActive: boolean = true;
 
   uid: any;
-  cambiotab: boolean = false;
   settings: any;
   regresarLabel: string;
 
   source: LocalDataSource = new LocalDataSource();
+  cambiotab: boolean;
   constructor(
     private translate: TranslateService,
     private popUpManager: PopUpManager,
     // tslint:disable-next-line
     private rqManager: RequestManager,
   ) {
+    this.cambiotab = false;
+    this.crudcambiotab.emit(false);
     this.auxcambiotab.emit(false);
     // console.log('constructor');
   }
@@ -60,6 +68,34 @@ export class ListEntityComponent implements OnInit {
     this.translate.onLangChange.subscribe((event: LangChangeEvent) => {
       this.cargarCampos();
     });
+    this.filtrarLista();
+  }
+  ngOnChanges(changes) {
+    if (changes['paramsFieldsName'] && changes['paramsFieldsName'].currentValue) {
+      this.paramsFieldsName = changes['paramsFieldsName'].currentValue;
+      this.loadData();
+    }
+
+  }
+  filtrarLista() {
+    this.source.onChanged().subscribe((change) => {
+
+      if (change.action === 'filter') {
+        /*        console.info(change);
+               console.info(change.filter.filters); */
+        change.filter.filters.map((item) => {
+          if (item.field === 'Vigencia' &&
+            (item.search.length === 4 || item.search === '0')) {
+            this.paramsFieldsName = { Vigencia: item.search, UnidadEjecutora: 1 };
+            this.loadData();
+          }
+        });
+
+        // Do whatever you want with the filter event
+
+      }
+    });
+
   }
   cargarCampos() {
     if (this.listSettings !== undefined) {
@@ -71,7 +107,7 @@ export class ListEntityComponent implements OnInit {
           edit: false,
           delete: false,
           custom: [{ name: 'edit', title: '<i class="nb-edit"></i>' }, { name: 'delete', title: '<i class="nb-trash"></i>' }],
-          position: 'left'
+          position: 'right'
         },
         add: {
           addButtonContent: '<i class="nb-plus"></i>',
@@ -90,7 +126,7 @@ export class ListEntityComponent implements OnInit {
 
   loadData(): void {
 
-    this.loadDataFunction('').subscribe(res => {
+    this.loadDataFunction('', this.paramsFieldsName ? this.paramsFieldsName : '').subscribe(res => {
       if (res !== null) {
         const data = <Array<any>>res;
         this.source.load(data);
@@ -99,15 +135,30 @@ export class ListEntityComponent implements OnInit {
       }
     });
   }
+  IsFuente(event) {
+    if (event.data.ValorInicial !== undefined) {
+      this.formEntity.campos[this.getIndexForm('Codigo')].deshabilitar = true;
+      this.paramsFieldsName['Vigencia'] = event.data.Vigencia;
+      if (event.data.Vigencia === 'sin vigencia asignada') {
+        this.paramsFieldsName['Vigencia'] = '0';
+      }
+    }
+  }
   onEdit(event): void {
     console.info(event);
     this.uid = event.data[this.uuidReadField];
+    this.IsFuente(event);
     this.activetab('crud');
+  }
+
+  emitItemSelected(event) {
+    this.infooutput.emit(event.data);
   }
 
   onAddOther(event): void {
     console.info(event);
-    this.activetab('other');
+    this.infooutput.emit(event.data);
+    this.activetab(event.action);
   }
   onCustom(event): void {
     switch (event.action) {
@@ -122,13 +173,19 @@ export class ListEntityComponent implements OnInit {
         this.onAddOther(event);
         break;
       default:
+        this.onAddOther(event);
         break;
     }
   }
 
   onCreate(event): void {
-    this.uid = null;
-    this.activetab('crud');
+    if (!this.externalCreate) {
+      this.formEntity.campos[this.getIndexForm('Codigo')].deshabilitar = false;
+      this.uid = null;
+      this.activetab('crud');
+    } else {
+      this.activetab('external-create');
+    }
   }
 
   onDelete(event): void {
@@ -142,12 +199,18 @@ export class ListEntityComponent implements OnInit {
     };
     Swal.fire(opt).then(willDelete => {
       if (willDelete.value) {
-        this.deleteDataFunction(event.data[this.uuidDeleteField]).subscribe(res => {
-          if (res !== null) {
+        this.deleteDataFunction(event.data[this.uuidDeleteField], this.paramsFieldsName ? this.paramsFieldsName : '').subscribe(res => {
+          if (res['Type'] === 'error') {
+            if ( res['Message']) {
+              this.popUpManager.showErrorAlert(res['Message']);
+            } else {
+              this.popUpManager.showErrorAlert(res['Body']);
+            }
+          } else {
             this.loadData();
             this.popUpManager.showSuccessAlert(
               this.translate.instant(this.deleteConfirmMessage)
-            );
+              );
           }
         });
       }
@@ -157,8 +220,13 @@ export class ListEntityComponent implements OnInit {
   activetab(tab): void {
     if (tab === 'crud') {
       this.cambiotab = !this.cambiotab;
+      this.crudcambiotab.emit(this.cambiotab);
     } else if (tab === 'other') {
       this.auxcambiotab.emit(true);
+      this.externalTabActivator.emit(tab);
+    } else {
+      this.externalTabActive = true;
+      this.externalTabActivator.emit(tab);
     }
   }
 
@@ -166,9 +234,11 @@ export class ListEntityComponent implements OnInit {
 
   selectTab(event): void {
     if (event.tabTitle === this.translate.instant('GLOBAL.lista')) {
+      this.crudcambiotab.emit(false);
       this.cambiotab = false;
     } else {
       this.cambiotab = true;
+      this.crudcambiotab.emit(true);
     }
   }
 
@@ -176,11 +246,20 @@ export class ListEntityComponent implements OnInit {
     if (event) {
       this.loadData();
       this.cambiotab = !this.cambiotab;
+      this.crudcambiotab.emit(this.cambiotab);
     }
   }
   itemselec(event): void {
-    // console.log('afssaf');
+    // console.info(event);
   }
-
+  getIndexForm(nombre: String): number {
+    for (let index = 0; index < this.formEntity.campos.length; index++) {
+      const element = this.formEntity.campos[index];
+      if (element.nombre === nombre) {
+        return index;
+      }
+    }
+    return 0;
+  }
 
 }
