@@ -157,12 +157,11 @@ export class DinamicformComponent implements OnInit, OnChanges {
     if (c.requerido && ((c.valor === '' && c.etiqueta !== 'file') || c.valor === null || c.valor === undefined ||
       (JSON.stringify(c.valor) === '{}' && c.etiqueta !== 'file') || JSON.stringify(c.valor) === '[]')
       || ((c.etiqueta === 'file' && c.valor.name === undefined) && (c.etiqueta === 'file' && c.urlTemp === undefined))) {
-      if (c.prefix) {
-        c.alerta = '** Este patrón no es aceptado';
-      } else {
-        c.alerta = '** Debe llenar este campo';
-
-      }
+        if (c.prefix) {
+          c.alerta = this.showErrorPatternMessage(c.pattern);
+        } else {
+          c.alerta = '** Debe llenar este campo';
+        }
       c.clase = 'form-control form-control-danger';
       return false;
     }
@@ -217,6 +216,15 @@ export class DinamicformComponent implements OnInit, OnChanges {
     this.normalform.campos.forEach(d => {
       d.valor = null;
     });
+  }
+
+  showErrorPatternMessage(pattern) {
+    switch (pattern) {
+      case '^[1-9]{1,9}':
+        return '** Formato Incorrecto: Solo se permite ingresar números';
+      default:
+        return '** Formato Incorrecto';
+    }
   }
 
   validForm() {
