@@ -1,17 +1,52 @@
 import { Component } from '@angular/core';
+import { MatIconRegistry } from '@angular/material/icon';
+import { DomSanitizer } from '@angular/platform-browser';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'ngx-footer',
   styleUrls: ['./footer.component.scss'],
-  template: `
-    <span class="created-by">Created with ♥ by <b><a href="https://akveo.com" target="_blank">Akveo</a></b> 2019</span>
-    <div class="socials">
-      <a href="#" target="_blank" class="ion ion-social-github"></a>
-      <a href="#" target="_blank" class="ion ion-social-facebook"></a>
-      <a href="#" target="_blank" class="ion ion-social-twitter"></a>
-      <a href="#" target="_blank" class="ion ion-social-linkedin"></a>
-    </div>
-  `,
+  templateUrl: './footer.component.html',
 })
+
 export class FooterComponent {
+  constructor(
+    private matIconRegistry: MatIconRegistry,
+    private translate: TranslateService,
+    private domSanitizer: DomSanitizer,
+    ) {
+    this.matIconRegistry.addSvgIcon(
+      'clock-outline',
+      this.domSanitizer.bypassSecurityTrustResourceUrl(
+        'presupuesto/../assets/images/clock-outline.svg'
+        )
+      );
+    this.matIconRegistry.addSvgIcon(
+      'globe-outline',
+      this.domSanitizer.bypassSecurityTrustResourceUrl(
+        'presupuesto/../assets/images/globe-outline.svg'
+        )
+      );
+    this.matIconRegistry.addSvgIcon(
+      'at-outline',
+      this.domSanitizer.bypassSecurityTrustResourceUrl(
+        'presupuesto/../assets/images/at-outline.svg'
+        )
+      );
+    this.matIconRegistry.addSvgIcon(
+      'phone-outline',
+      this.domSanitizer.bypassSecurityTrustResourceUrl(
+        'presupuesto/../assets/images/phone-outline.svg'
+        )
+      );
+
+    this.matIconRegistry.addSvgIcon(
+      'pin-outline',
+      this.domSanitizer.bypassSecurityTrustResourceUrl(
+        'presupuesto/../assets/images/pin-outline.svg'
+      )
+    );
+    translate.use(translate.getBrowserLang());
+  }
 }
+
