@@ -1,4 +1,7 @@
 import { Component } from '@angular/core';
+import { NbSidebarService } from '@nebular/theme';
+import { LayoutService } from '../../../@core/utils';
+
 
 @Component({
   selector: 'ngx-one-column-layout',
@@ -9,7 +12,7 @@ import { Component } from '@angular/core';
         <ngx-header></ngx-header>
       </nb-layout-header>
 
-      <nb-sidebar class="menu-sidebar" tag="menu-sidebar" responsive>
+      <nb-sidebar class="menu-sidebar" (click)="toggleSidebar($event)" tag="menu-sidebar" responsive>
         <ng-content select="nb-menu"></ng-content>
       </nb-sidebar>
 
@@ -23,4 +26,22 @@ import { Component } from '@angular/core';
     </nb-layout>
   `,
 })
-export class OneColumnLayoutComponent {}
+export class OneColumnLayoutComponent {
+
+  constructor(
+    private sidebarService: NbSidebarService,
+    private layoutService: LayoutService,
+  ) {
+
+  }
+
+  toggleSidebar(event): boolean {
+    if (event.target.nextSibling === null &&
+      (event.target.className === 'menu-title' || event.target.lastElementChild.className === 'menu-title')) {
+      this.sidebarService.toggle(true, 'menu-sidebar');
+      this.layoutService.changeLayoutSize();
+      return false;
+    }
+    return true;
+  }
+}
