@@ -1,8 +1,14 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { Rubro } from '../../../@core/data/models/rubro';
 import { ApropiacionHelper } from '../../../@core/helpers/apropiaciones/apropiacionHelper';
+import { FuenteHelper } from '../../../@core/helpers/fuentes/fuenteHelper';
 import { PopUpManager } from '../../../@core/managers/popUpManager';
 import { ArbolApropiacion } from '../../../@core/data/models/arbol_apropiacion';
+import { CommonHelper } from '../../../@core/helpers/commonHelper';
+import { DependenciaHelper } from '../../../@core/helpers/oikos/dependenciaHelper';
+import { registerLocaleData } from '@angular/common';
+import locales from '@angular/common/locales/es-CO';
+registerLocaleData(locales, 'co');
 
 @Component({
   selector: 'ngx-apropiaciones',
@@ -22,22 +28,36 @@ export class ApropiacionesComponent implements OnInit {
   vigenciaSel: any;
   clean = false;
   opcion: string;
-  VigenciaActual = '2020';
+  VigenciaActual = '2020'; // TODO: traer del endpoint vigencia_actual
   optionView: string;
+  productos: boolean = false;
+  listaProductosAsignados = [];
   vigencias: any[] = [
     { vigencia: 2020 },
     { vigencia: 2019 },
     { vigencia: 2018 },
     { vigencia: 2017 },
     { vigencia: 2016 },
-  ];
+  ];  // TODO: traer del endpoint vigencias_afuncional
   balanceado: boolean;
+  allApproved: boolean;
+  AreaFuncional: string;
+  CentroGestor: string;
+  planAdquisicionesRubro: any;
+  paramsFieldsName: object;
+  totalValorActividades: number;
+  diferenciaActividadApropiacion: number;
+  totalValorFuentes: number;
+  diferenciaFuentesApropiacion: number;
 
   constructor(
     private apHelper: ApropiacionHelper,
+    private fuenteHelper: FuenteHelper,
+    private commonHelper: CommonHelper,
     private popManager: PopUpManager,
+    private dependenciaHelper: DependenciaHelper,
   ) {
-    this.vigenciaSel = '2020';
+    this.vigenciaSel = '2020';    // TODO: traer del endpoint vigencia_actual +1
     this.optionView = 'Apropiaciones';
 
     this.rubroSeleccionado = {
@@ -54,7 +74,7 @@ export class ApropiacionesComponent implements OnInit {
 
     this.apropiacionData = {
       Vigencia: 0,
-      ApropiacionInicial: 0,
+      ValorInicial: 0,
       ApropiacionAnterior: 0,
       Estado: '',
       Rubro: <Rubro>{},
@@ -64,32 +84,45 @@ export class ApropiacionesComponent implements OnInit {
       UnidadEjecutora: '',
       Padre: '',
       Hijos: [],
+      Productos: []
     };
 
   }
 
 
   ngOnInit() {
-
+    this.commonHelper.geCurrentVigencia(1).subscribe(res => {
+      if (res) {
+        this.vigenciaSel = res + '';
+      }
+    });
+    this.paramsFieldsName = { Vigencia: this.vigenciaSel, UnidadEjecutora: 1 };
   }
 
   receiveMessage($event) {
     if ($event.Hijos.length === 0) {
       this.isLeaf = true;
-      this.rubroSeleccionado = <Rubro>$event;
-      // console.info(this.rubroSeleccionado);
+      this.rubroSeleccionado = <ArbolApropiacion>$event;
       this.rubroSeleccionado.Id = parseInt(this.rubroSeleccionado.Id, 0);
       this.rubroSeleccionado.Nombre = this.rubroSeleccionado.Nombre;
+      this.CentroGestor = '230';
+      this.AreaFuncional = '0' + this.rubroSeleccionado.UnidadEjecutora + '-Rector';
       this.rubroSeleccionado.UnidadEjecutora = parseInt(
         this.rubroSeleccionado.UnidadEjecutora,
         0,
       );
-      this.rubroSeleccionado.ApropiacionInicial = parseInt(this.rubroSeleccionado.ApropiacionInicial, 0);
+      this.rubroSeleccionado.ValorInicial = this.rubroSeleccionado.ValorInicial ? parseInt(this.rubroSeleccionado.ValorInicial, 0) : 0;
+
+      this.valorApropiacion = this.rubroSeleccionado.ValorInicial;
+      if (this.rubroSeleccionado.Estado === 'registrada') {
+        this.productos = true;
+      }
+      this.listaProductosAsignados = this.rubroSeleccionado.Productos;
     } else {
       this.isLeaf = false;
+      this.productos = false;
     }
   }
-
 
   aprobarApropiacion() {
     this.popManager.showAlert('warning', 'Aprobar Apropiación', 'esta seguro?')
@@ -98,7 +131,7 @@ export class ApropiacionesComponent implements OnInit {
           this.apHelper.apropiacionApprove({ UnidadEjecutora: '1', Vigencia: this.vigenciaSel }).subscribe((res) => {
             if (res) {
               this.popManager.showSuccessAlert('Aprobación exitosa para la apropiación ' + this.vigenciaSel);
-              this.cleanForm();
+              this.cleanForm(true);
               this.eventChange.emit(true);
             }
           });
@@ -109,7 +142,7 @@ export class ApropiacionesComponent implements OnInit {
   }
 
 
-  cleanForm() {
+  cleanForm(full?: boolean) {
     this.clean = !this.clean;
     this.rubroSeleccionado = {
       Id: 0,
@@ -124,7 +157,9 @@ export class ApropiacionesComponent implements OnInit {
     };
     this.apropiacionData = <ArbolApropiacion>{};
     this.valorApropiacion = 0;
-
+    if (full) {
+      this.isLeaf = false;
+    }
   }
 
   preAsignarApropiacion() {
@@ -135,16 +170,15 @@ export class ApropiacionesComponent implements OnInit {
     this.apropiacionData.UnidadEjecutora = typeof this.rubroSeleccionado.UnidadEjecutora === 'undefined' ? undefined : this.rubroSeleccionado.UnidadEjecutora;
     this.apropiacionData.Padre = typeof this.rubroSeleccionado.Padre === 'undefined' ? undefined : this.rubroSeleccionado.Padre;
     this.apropiacionData.Hijos = typeof this.rubroSeleccionado.Hijos === 'undefined' ? undefined : this.rubroSeleccionado.Hijos;
-    this.apropiacionData.ApropiacionInicial = typeof this.valorApropiacion === 'undefined' ? undefined : this.valorApropiacion;
-    this.apropiacionData.ApropiacionAnterior = typeof this.rubroSeleccionado.ApropiacionInicial === 'undefined' ? undefined : this.rubroSeleccionado.ApropiacionInicial;
+    this.apropiacionData.ValorInicial = typeof this.valorApropiacion === 'undefined' ? undefined : this.valorApropiacion;
+    this.apropiacionData.ApropiacionAnterior = typeof this.rubroSeleccionado.ValorInicial === 'undefined' ? 0 : this.rubroSeleccionado.ValorInicial;
     this.apropiacionData.Estado = 'registrada'; // Estado preasignado
 
-    console.table(this.apropiacionData);
     if (this.vigenciaSel !== undefined) {
       this.apHelper.apropiacionRegister(this.apropiacionData).subscribe((res) => {
         if (res) {
           this.popManager.showSuccessAlert('Se registro la preasignación de apropiación correctamente!');
-          this.cleanForm();
+          // this.cleanForm();
           this.eventChange.emit(true);
         }
       });
@@ -158,11 +192,14 @@ export class ApropiacionesComponent implements OnInit {
   onSelect(selectedItem: any) {
     this.vigenciaSel = selectedItem;
     // this.eventChange.emit(true);
-    console.info(this.vigenciaSel);
   }
 
-  checkComprobacion(event: boolean) {
-    this.balanceado = event;
+  checkComprobacion(event: { balanceado: boolean, approved: boolean }) {
+    this.balanceado = event.balanceado;
+    this.allApproved = event.approved;
+  }
+  cambioProductosAsignados(productosAsignados: any[]) {
+    this.listaProductosAsignados = productosAsignados;
   }
 
 }
