@@ -3,21 +3,21 @@ export let DETALLE_MODIFICACION_FORM = {
     tipo_formulario: 'mini',
     alertas: true,
     modelo: 'detalleModificacion',
-    btn: 'GLOBAL.siguiente_abr',
+    btn: 'GLOBAL.siguiente',
     campos: [
         {
             etiqueta: 'input',
-            claseGrid: 'mx-auto col-lg-4 col-md-4 col-sm-4 col-xs-4',
-            nombre: 'nDocumento',
+            claseGrid: 'mx-auto col-lg-8 col-md-8 col-sm-8 col-xs-8',
+            nombre: 'NumeroDocumento',
             label_i18n: 'numero_documento',
             placeholder_i18n: 'numero_documento',
             requerido: true,
-            tipo: 'number',
+            tipo: 'text',
         },
         {
             etiqueta: 'select',
-            claseGrid: 'mx-auto col-lg-4 col-md-4 col-sm-4 col-xs-4',
-            nombre: 'tipoDocumento',
+            claseGrid: 'mx-auto col-lg-8 col-md-8 col-sm-8 col-xs-8',
+            nombre: 'TipoDocumento',
             label_i18n: 'tipo_documento',
             placeholder_i18n: 'tipo_documento',
             requerido: true,
@@ -26,16 +26,25 @@ export let DETALLE_MODIFICACION_FORM = {
         },
         {
             etiqueta: 'mat-date',
-            claseGrid: 'mx-auto col-lg-4 col-md-4 col-sm-4 col-xs-4',
-            nombre: 'fDocumento',
+            claseGrid: 'mx-auto col-lg-8 col-md-8 col-sm-8 col-xs-8',
+            nombre: 'FechaDocumento',
             label_i18n: 'fecha_documento',
             placeholder_i18n: 'fecha_documento',
             requerido: true,
         },
         {
+            etiqueta: 'input',
+            claseGrid: 'mx-auto col-lg-8 col-md-8 col-sm-8 col-xs-8',
+            nombre: 'OrganismoEmisor',
+            label_i18n: 'organismo_emisor',
+            placeholder_i18n: 'organismo_emisor',
+            requerido: true,
+            tipo: 'text',
+        },
+        {
             etiqueta: 'textarea',
-            claseGrid: 'mx-auto col-lg-7 col-md-7 col-sm-7 col-xs-7',
-            nombre: 'descripcion',
+            claseGrid: 'mx-auto col-lg-8 col-md-8 col-sm-8 col-xs-8',
+            nombre: 'Descripcion',
             label_i18n: 'descripcion',
             placeholder_i18n: 'descripcion',
             requerido: true,
