@@ -6,12 +6,19 @@ import { RubrosComponent } from './rubros/rubros.component';
 import { GestionApropiacionesComponent } from './gestion-apropiaciones/gestion-apropiaciones.component';
 import { ListProductoComponent } from './productos/list-producto.component';
 import { ListFuenteComponent } from './fuentes/list-fuente/list-fuente.component';
+import { ListSolicitudCdpComponent } from './cdp/list-solicitud-cdp/list-solicitud-cdp.component';
+import { ListCdpComponent } from './cdp/list-cdp/list-cdp.component';
 import { ModificacionApropiacionComponent } from './modificacion-apropiacion/modificacion-apropiacion.component';
 import { SetModificacionApropiacionComponent } from './modificacion-apropiacion/set-modificacion-apropiacion/set-modificacion-apropiacion.component';
 import { ShowModificationAfectationComponent } from './modificacion-apropiacion/show-modification-afectation/show-modification-afectation.component';
 import { ShowModificationResumeComponent } from './modificacion-apropiacion/show-modification-resume/show-modification-resume.component';
-
-
+import { SolicitudCrpComponent } from './crp/solicitud-crp/solicitud-crp.component';
+import { ListSolicitudCrpComponent } from './crp/list-solicitud-crp/list-solicitud-crp.component';
+import { ListCrpComponent } from './crp/list-crp/list-crp.component';
+import { ListModificacionApropiacionComponent } from './consulta-modificacion-apropiacion/list-modificacion-apropiacion.component';
+import { ShowModificationApropiacionDataComponent } from './consulta-modificacion-apropiacion/show-modificacion-apropiacion/show-modificacion-apropiacion.component';
+import { ListarVigenciaComponent } from './vigencia/listar-vigencia/listar-vigencia.component';
+import { CierreVigenciaComponent } from './gestion-vigencias/cierre-vigencia/cierre-vigencia.component';
 
 const routes: Routes = [{
   path: 'arbol',
@@ -30,6 +37,10 @@ const routes: Routes = [{
   component: GestionApropiacionesComponent,
 },
 {
+  path: 'apropiaciones/:vista',
+  component: GestionApropiacionesComponent,
+},
+{
   path: 'productos',
   component: ListProductoComponent,
 },
@@ -38,14 +49,46 @@ const routes: Routes = [{
   component: ListFuenteComponent,
 },
 {
+  path: 'solicitudcdp',
+  component: ListSolicitudCdpComponent,
+},
+ {
   path: 'modificacion-apropiacion',
-  component: ModificacionApropiacionComponent,
+  component: ListModificacionApropiacionComponent,
+},
+{
+  path: 'cdp',
+  component: ListCdpComponent,
+},
+{
+  path: 'crp',
+  component: ListCrpComponent,
 },
 {
   path: '',
   redirectTo: 'rubros',
   pathMatch: 'full',
-}
+},
+{
+  path: 'form-solicitudcrp',
+  component: SolicitudCrpComponent,
+},
+{
+  path: 'solicitudcrp',
+  component: ListSolicitudCrpComponent,
+},
+{
+  path: 'list-modificacion-apropiacion',
+  component: ListModificacionApropiacionComponent,
+},
+{
+  path: 'listar-vigencias',
+  component: ListarVigenciaComponent,
+},
+{
+  path: 'cierre-vigencia',
+  component: CierreVigenciaComponent,
+},
 
 ];
 
@@ -65,4 +108,7 @@ export const routedComponents = [
   SetModificacionApropiacionComponent,
   ShowModificationAfectationComponent,
   ShowModificationResumeComponent,
+  ShowModificationApropiacionDataComponent,
+  ListModificacionApropiacionComponent,
+  CierreVigenciaComponent,
 ];
