@@ -10,6 +10,8 @@ import {
   NbTabsetModule,
   NbStepperModule,
   NbCardModule,
+  NbTooltipModule,
+  NbRadioModule,
   NbCheckboxModule } from '@nebular/theme';
 
 import { ProductosRubroComponent } from './rubros/productos-rubro/productos-rubro.component';
@@ -23,7 +25,25 @@ import { ToasterModule, ToasterService } from 'angular2-toaster';
 import { ConfiguracionService } from '../../@core/data/configuracion.service';
 import { MatStepperModule } from '@angular/material';
 import { DependenciasComponent } from './dependencias/dependencias.component';
-
+import { ListSolicitudCdpComponent } from './cdp/list-solicitud-cdp/list-solicitud-cdp.component';
+import { VerSolicitudCdpComponent } from './cdp/ver-solicitud-cdp/ver-solicitud-cdp.component';
+import { ListCdpComponent } from './cdp/list-cdp/list-cdp.component';
+import { ListCrpComponent } from './crp/list-crp/list-crp.component';
+import { ExpedirCrpComponent } from './crp/expedir-crp/expedir-crp.component';
+import { ListSolicitudCrpComponent } from './crp/list-solicitud-crp/list-solicitud-crp.component';
+import { VerSolicitudCrpComponent } from './crp/ver-solicitud-crp/ver-solicitud-crp.component';
+import { SolicitudCrpComponent } from './crp/solicitud-crp/solicitud-crp.component';
+import { CurrencyMaskModule } from 'ng2-currency-mask';
+import { ListModificacionApropiacionComponent } from './consulta-modificacion-apropiacion/list-modificacion-apropiacion.component';
+import { RubrosFuenteComponent } from './fuentes/rubros-fuente/rubros-fuente.component';
+import { AnulacionDocumentoComponent } from './anulacion-documento-pres/anulacion-documento.component';
+import { SetModificacionFuenteComponent } from './modificacion-apropiacion/set-modificacion-fuente/set-modificacion-fuente.component';
+import { ShowModificationFuenteComponent } from './modificacion-apropiacion/show-modification-fuente/show-modification-fuente.component';
+import { CierreVigenciaComponent } from './gestion-vigencias/cierre-vigencia/cierre-vigencia.component';
+import { PlanAdquisicionesRubroComponent } from './rubros/plan-adquisiciones-rubro/plan-adquisiciones-rubro.component';
+import { ListAnulacionDocumentoComponent } from './list-anulacion-documento-pres/list-anulacion-documento.component';
+import { ListarVigenciaComponent } from './vigencia/listar-vigencia/listar-vigencia.component';
+import { ShowVigenciaComponent } from './vigencia/show-vigencia/show-vigencia.component';
 
 
 @NgModule({
@@ -36,6 +56,24 @@ import { DependenciasComponent } from './dependencias/dependencias.component';
     ListFuenteComponent,
     ListProductoComponent,
     DependenciasComponent,
+    ListCrpComponent,
+    ListSolicitudCdpComponent,
+    ListSolicitudCrpComponent,
+    VerSolicitudCdpComponent,
+    SolicitudCrpComponent,
+    ListCdpComponent,
+    VerSolicitudCrpComponent,
+    ListModificacionApropiacionComponent,
+    RubrosFuenteComponent,
+    AnulacionDocumentoComponent,
+    SetModificacionFuenteComponent,
+    ShowModificationFuenteComponent,
+    CierreVigenciaComponent,
+    PlanAdquisicionesRubroComponent,
+    ListAnulacionDocumentoComponent,
+    ListarVigenciaComponent,
+    ShowVigenciaComponent,
+    ExpedirCrpComponent
   ],
   providers: [
     ConfiguracionService,
@@ -47,7 +85,9 @@ import { DependenciasComponent } from './dependencias/dependencias.component';
     CommonModule,
     PlanCuentasRoutingModule,
     NbTreeGridModule,
+    NbTooltipModule,
     NbSelectModule,
+    NbRadioModule,
     NbAlertModule,
     NbCheckboxModule,
     NbTabsetModule,
@@ -56,6 +96,7 @@ import { DependenciasComponent } from './dependencias/dependencias.component';
     Ng2SmartTableModule,
     ToasterModule,
     MatStepperModule,
+    CurrencyMaskModule,
   ],
   exports: [
   ],
