@@ -4,23 +4,42 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 export const environment = {
-  production: false,
-  NUXEO: {
-    PATH: 'https://documental.udistrital.edu.co/nuxeo/',
-  },
-  WSO2_SERVICE: 'https://autenticacion.portaloas.udistrital.edu.co/apioas/',
-  PLAN_CUENTAS_MONGO_SERVICE: 'http://localhost:8082/v1/',
-  OIKOS_SERVICE: 'https://autenticacion.portaloas.udistrital.edu.co/apioas/oikos_api/v1/',
-  CONFIGURACION_SERVICE: 'https://autenticacion.portaloas.udistrital.edu.co/apioas/configuracion_crud_api/v1/',
-  CONF_MENU_SERVICE: 'https://autenticacion.portaloas.udistrital.edu.co/apioas/configuracion_crud_api/v1/menu_opcion_padre/ArbolMenus/',
-  TOKEN: {
-    AUTORIZATION_URL: 'https://autenticacion.portaloas.udistrital.edu.co/oauth2/authorize',
-    CLIENTE_ID: 'e36v1MPQk2jbz9KM4SmKhk8Cyw0a',
-    RESPONSE_TYPE: 'id_token token',
-    SCOPE: 'openid email role',
-    REDIRECT_URL: 'http://localhost:4200/',
-    SIGN_OUT_URL: 'https://autenticacion.portaloas.udistrital.edu.co/oidc/logout',
-    SIGN_OUT_REDIRECT_URL: 'http://localhost:4200/',
-  },
+
+    production: false,
+    NUXEO: {
+        PATH: 'https://documental.udistrital.edu.co/nuxeo/',
+    },
+    WSO2_SERVICE: 'https://autenticacion.portaloas.udistrital.edu.co/apioas/',
+    NECESIDADES_CRUD_SERVICE: "http://pruebasapi.intranetoas.udistrital.edu.co:8201/v1/",
+    PLAN_ADQUISICION_SERVICE: 'https://autenticacion.portaloas.udistrital.edu.co/apioas/bodega_jbpm/v1/',
+    // MOVIMIENTOS_CRUD_SERVICE: 'http://localhost:8085/v1/',
+    ADMINISTRATIVA_PRUEBAS_SERVICE: 'http://pruebasapi.intranetoas.udistrital.edu.co:8104/v1/',
+    PLAN_CUENTAS_MID_SERVICE: 'http://localhost:8084/v1/',
+    // PLAN_CUENTAS_MID_SERVICE: 'http://localhost:8084/v1/',
+    PLAN_CUENTAS_MONGO_SERVICE: 'http://localhost:8082/v1/',
+    // PLAN_CUENTAS_MONGO_SERVICE: 'http://pruebasapi.intranetoas.udistrital.edu.co:8203/v1/',
+    // PLAN_CUENTAS_MID_SERVICE: 'http://pruebasapi.intranetoas.udistrital.edu.co:8204/v1/',
+    ADMINISTRATIVA_SERVICE: 'http://pruebasapi.intranetoas.udistrital.edu.co:8090/v1/',
+    ADMINISTRATIVA_JBPM_SERVICE: 'https://autenticacion.portaloas.udistrital.edu.co/apioas/administrativa_jbpm/v1/',
+    // NECESIDADES_CRUD_SERVICE: 'https://autenticacion.udistrital.edu.co/apioas/necesidades_crud/v1/',
+    // MOVIMIENTOS_CRUD_SERVICE: 'https://autenticacion.portaloas.udistrital.edu.co/apioas/movimientos_crud/v1/',
+    MOVIMIENTOS_CRUD_SERVICE: 'http://localhost:8085/v1/',
+    // MOVIMIENTOS_CRUD_SERVICE: 'https://autenticacion.udistrital.edu.co/apioas/movimientos_crud/v1/',
+    // ADMINISTRATIVA_SERVICE: 'http://pruebasapi.intranetoas.udistrital.edu.co:8201/v1/',
+    OIKOS_SERVICE: 'https://autenticacion.portaloas.udistrital.edu.co/apioas/oikos_crud_api/v2/',
+    CORE_SERVICE: 'http://pruebasapi.intranetoas.udistrital.edu.co:8092/v1/',
+    CORE_AMAZON_SERVICE: "https://autenticacion.portaloas.udistrital.edu.co/apioas/core_amazon_crud/v1/",
+    CONFIGURACION_SERVICE: 'https://autenticacion.portaloas.udistrital.edu.co/apioas/configuracion_crud_api/v1/',
+    CONF_MENU_SERVICE: 'https://autenticacion.portaloas.udistrital.edu.co/apioas/configuracion_crud_api/v1/menu_opcion_padre/ArbolMenus/',
+    TOKEN: {
+        AUTORIZATION_URL: 'https://autenticacion.portaloas.udistrital.edu.co/oauth2/authorize',
+        CLIENTE_ID: 'e36v1MPQk2jbz9KM4SmKhk8Cyw0a',
+        RESPONSE_TYPE: 'id_token token',
+        SCOPE: 'openid email role',
+        REDIRECT_URL: 'http://localhost:4200/',
+        SIGN_OUT_URL: 'https://autenticacion.portaloas.udistrital.edu.co/oidc/logout',
+        SIGN_OUT_REDIRECT_URL: 'http://localhost:4200/',
+    },
+
 
 };
